+++ src/tinflate.c
@@ -29,7 +29,7 @@
 #include <limits.h>
 
 #if defined(ULONG_MAX) && (ULONG_MAX) < 0xFFFFFFFFUL
-#  error "tinf requires unsigned long to be at least 32-bit"
+#  error "tinf requires uint32_t to be at least 32-bit"
 #endif
 
 /* -- Internal data structures -- */
@@ -37,15 +37,15 @@
 struct tinf_tree {
     unsigned short counts[16]; /* Number of codes with a given length */
     unsigned short symbols[288]; /* Symbols sorted by code */
-    long max_sym;
+    uint32_t max_sym;
 };
 
 struct tinf_data {
     const unsigned char *source;
     const unsigned char *source_end;
-    unsigned long tag;
-    long bitcount;
-    long overflow;
+    uint32_t tag;
+    uint32_t bitcount;
+    uint32_t overflow;
 
     unsigned char *dest_start;
     unsigned char *dest;
@@ -57,16 +57,16 @@ struct tinf_data {
 
 /* -- Utility functions -- */
 
-static unsigned long read_le16(const unsigned char *p)
+static uint32_t read_le16(const unsigned char *p)
 {
-    return ((unsigned long) p[0])
-         | ((unsigned long) p[1] << 8);
+    return ((uint32_t) p[0])
+         | ((uint32_t) p[1] << 8);
 }
 
 /* Build fixed Huffman trees */
 static void tinf_build_fixed_trees(struct tinf_tree *lt, struct tinf_tree *dt)
 {
-    long i;
+    uint32_t i;
 
     /* Build fixed literal/length tree */
     for (i = 0; i < 16; ++i) {
@@ -107,11 +107,11 @@ static void tinf_build_fixed_trees(struct tinf_tree *lt, struct tinf_tree *dt)
 }
 
 /* Given an array of code lengths, build a tree */
-static long tinf_build_tree(struct tinf_tree *t, const unsigned char *lengths,
-                           unsigned long num)
+static uint32_t tinf_build_tree(struct tinf_tree *t, const unsigned char *lengths,
+                           uint32_t num)
 {
     unsigned short offs[16];
-    unsigned long i, num_codes, available;
+    uint32_t i, num_codes, available;
 
     assert(num <= 288);
 
@@ -133,7 +133,7 @@ static long tinf_build_tree(struct tinf_tree *t, const unsigned char *lengths,
 
     /* Compute offset table for distribution sort */
     for (available = 1, num_codes = 0, i = 0; i < 16; ++i) {
-        unsigned long used = t->counts[i];
+        uint32_t used = t->counts[i];
 
         /* Check length contains no more codes than available */
         if (used > available) {
@@ -175,14 +175,14 @@ static long tinf_build_tree(struct tinf_tree *t, const unsigned char *lengths,
 
 /* -- Decode functions -- */
 
-static void tinf_refill(struct tinf_data *d, long num)
+static void tinf_refill(struct tinf_data *d, uint32_t num)
 {
     assert(num >= 0 && num <= 32);
 
     /* Read bytes until at least num bits available */
     while (d->bitcount < num) {
         if (d->source != d->source_end) {
-            d->tag |= (unsigned long) *d->source++ << d->bitcount;
+            d->tag |= (uint32_t) *d->source++ << d->bitcount;
         }
         else {
             d->overflow = 1;
@@ -193,9 +193,9 @@ static void tinf_refill(struct tinf_data *d, long num)
     assert(d->bitcount <= 32);
 }
 
-static unsigned long tinf_getbits_no_refill(struct tinf_data *d, long num)
+static uint32_t tinf_getbits_no_refill(struct tinf_data *d, uint32_t num)
 {
-    unsigned long bits;
+    uint32_t bits;
 
     assert(num >= 0 && num <= d->bitcount);
 
@@ -210,23 +210,23 @@ static unsigned long tinf_getbits_no_refill(struct tinf_data *d, long num)
 }
 
 /* Get num bits from source stream */
-static unsigned long tinf_getbits(struct tinf_data *d, long num)
+static uint32_t tinf_getbits(struct tinf_data *d, uint32_t num)
 {
     tinf_refill(d, num);
     return tinf_getbits_no_refill(d, num);
 }
 
 /* Read a num bit value from stream and add base */
-static unsigned long tinf_getbits_base(struct tinf_data *d, long num, long base)
+static uint32_t tinf_getbits_base(struct tinf_data *d, uint32_t num, uint32_t base)
 {
     return base + (num ? tinf_getbits(d, num) : 0);
 }
 
 /* Given a data stream and a tree, decode a symbol */
-static long tinf_decode_symbol(struct tinf_data *d, const struct tinf_tree *t)
+static uint32_t tinf_decode_symbol(struct tinf_data *d, const struct tinf_tree *t)
 {
-    long base = 0, offs = 0;
-    long len;
+    uint32_t base = 0, offs = 0;
+    uint32_t len;
 
     /*
      * Get more bits while code index is above number of codes
@@ -260,7 +260,7 @@ static long tinf_decode_symbol(struct tinf_data *d, const struct tinf_tree *t)
 }
 
 /* Given a data stream, decode dynamic trees from it */
-static long tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
+static uint32_t tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
                              struct tinf_tree *dt)
 {
     unsigned char lengths[288 + 32];
@@ -271,9 +271,9 @@ static long tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
         11,  4, 12, 3, 13, 2, 14, 1, 15
     };
 
-    unsigned long hlit, hdist, hclen;
-    unsigned long i, num, length;
-    long res;
+    uint32_t hlit, hdist, hclen;
+    uint32_t i, num, length;
+    uint32_t res;
 
     /* Get 5 bits HLIT (257-286) */
     hlit = tinf_getbits_base(d, 5, 257);
@@ -304,7 +304,7 @@ static long tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
     /* Read code lengths for code length alphabet */
     for (i = 0; i < hclen; ++i) {
         /* Get 3 bits code length (0-7) */
-        unsigned long clen = tinf_getbits(d, 3);
+        uint32_t clen = tinf_getbits(d, 3);
 
         lengths[clcidx[i]] = clen;
     }
@@ -323,7 +323,7 @@ static long tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
 
     /* Decode code lengths for the dynamic trees */
     for (num = 0; num < hlit + hdist; ) {
-        long sym = tinf_decode_symbol(d, lt);
+        uint32_t sym = tinf_decode_symbol(d, lt);
 
         if (sym > lt->max_sym) {
             return TINF_DATA_ERROR;
@@ -387,7 +387,7 @@ static long tinf_decode_trees(struct tinf_data *d, struct tinf_tree *lt,
 /* -- Block inflate functions -- */
 
 /* Given a stream and two trees, inflate a block of data */
-static long tinf_inflate_block_data(struct tinf_data *d, struct tinf_tree *lt,
+static uint32_t tinf_inflate_block_data(struct tinf_data *d, struct tinf_tree *lt,
                                    struct tinf_tree *dt)
 {
     /* Extra bits and base tables for length codes */
@@ -417,7 +417,7 @@ static long tinf_inflate_block_data(struct tinf_data *d, struct tinf_tree *lt,
     };
 
     for (;;) {
-        long sym = tinf_decode_symbol(d, lt);
+        uint32_t sym = tinf_decode_symbol(d, lt);
 
         /* Check for overflow in bit reader */
         if (d->overflow) {
@@ -431,8 +431,8 @@ static long tinf_inflate_block_data(struct tinf_data *d, struct tinf_tree *lt,
             *d->dest++ = sym;
         }
         else {
-            long length, dist, offs;
-            long i;
+            uint32_t length, dist, offs;
+            uint32_t i;
 
             /* Check for end of block */
             if (sym == 256) {
@@ -480,9 +480,9 @@ static long tinf_inflate_block_data(struct tinf_data *d, struct tinf_tree *lt,
 }
 
 /* Inflate an uncompressed block of data */
-static long tinf_inflate_uncompressed_block(struct tinf_data *d)
+static uint32_t tinf_inflate_uncompressed_block(struct tinf_data *d)
 {
-    unsigned long length, invlength;
+    uint32_t length, invlength;
 
     if (d->source_end - d->source < 4) {
         return TINF_DATA_ERROR;
@@ -522,7 +522,7 @@ static long tinf_inflate_uncompressed_block(struct tinf_data *d)
 }
 
 /* Inflate a block of data compressed with fixed Huffman trees */
-static long tinf_inflate_fixed_block(struct tinf_data *d)
+static uint32_t tinf_inflate_fixed_block(struct tinf_data *d)
 {
     /* Build fixed Huffman trees */
     tinf_build_fixed_trees(&d->ltree, &d->dtree);
@@ -532,10 +532,10 @@ static long tinf_inflate_fixed_block(struct tinf_data *d)
 }
 
 /* Inflate a block of data compressed with dynamic Huffman trees */
-static long tinf_inflate_dynamic_block(struct tinf_data *d)
+static uint32_t tinf_inflate_dynamic_block(struct tinf_data *d)
 {
     /* Decode trees from stream */
-    long res = tinf_decode_trees(d, &d->ltree, &d->dtree);
+    uint32_t res = tinf_decode_trees(d, &d->ltree, &d->dtree);
 
     if (res != TINF_OK) {
         return res;
@@ -554,11 +554,11 @@ void tinf_init(void)
 }
 
 /* Inflate stream from source to dest */
-long tinf_uncompress(void *dest, unsigned long *destLen,
-                    const void *source, unsigned long sourceLen)
+uint32_t tinf_uncompress(void *dest, uint32_t *destLen,
+                    const void *source, uint32_t sourceLen)
 {
     struct tinf_data d;
-    long bfinal;
+    uint32_t bfinal;
 
     /* Initialise data */
     d.source = (const unsigned char *) source;
@@ -572,8 +572,8 @@ long tinf_uncompress(void *dest, unsigned long *destLen,
     d.dest_end = d.dest + *destLen;
 
     do {
-        unsigned long btype;
-        long res;
+        uint32_t btype;
+        uint32_t res;
 
         /* Read final block flag */
         bfinal = tinf_getbits(&d, 1);
@@ -629,7 +629,7 @@ extern int
 LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
 {
     if (size > UINT_MAX / 2) { return 0; }
-    unsigned long destLen = sizeof(depacked);
+    uint32_t destLen = sizeof(depacked);
     tinf_uncompress(depacked, &destLen, data, size);
     return 0;
 }
